pyproject.toml
[project]
name = "xctl"
dynamic = ["version"]
description = "Xray Controller CLI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.0",
]

[project.scripts]
xctl = "src.cli.main:app"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "scm"
major_version_zero = true
update_changelog_on_bump = true
